Enable users to have their own lists and prevent /commands from being added to lists

+++ dbhelper.py
@@ -8,26 +8,28 @@ class DBHelper:
 
 	# Create table in DB called items with description column 
 	def setup(self):
-		stmt = "CREATE TABLE IF NOT EXISTS items (description text)"
+		print("creating table")
+		stmt = "CREATE TABLE IF NOT EXISTS items (description text, owner text)"
 		self.conn.execute(stmt)
 		self.conn.commit()
 
 	# Insert item text into DB 
-	def add_item(self, item_text):
-		stmt = "INSERT INTO items (description) VALUES (?)"
-		args = (item_text, )
+	def add_item(self, item_text, owner):
+		stmt = "INSERT INTO items (description, owner) VALUES (?. ?)"
+		args = (item_text, owner)
 		self.conn.execute(stmt, args)
 		self.conn.commit()
 
 	# Delete item text from DB 
 	def delete_item(self, item_text):
-		stmt = "DELETE FROM items WHERE description = (?)"
-		args = (item_text, )
+		stmt = "DELETE FROM items WHERE description = (?) AND owner = (?)"
+		args = (item_text, owner)
 		self.conn.execute(stmt, args)
 		self.conn.commit()
 
 	# Return a list of all items in DB 
 	# SQLite always returns data in tuple
-	def get_items(self):
-		stmt = "SELECT description FROM items"
-		return [x[0] for x in self.conn.execute(stmt)]
+	def get_items(self, owner):
+		stmt = "SELECT description FROM items WHERE owner = (?)"
+		args = (owner, )
+		return [x[0] for x in self.conn.execute(stmt, args)]
